+++ dimcli/__init__.py
@@ -39,7 +39,11 @@ except:
 
 
 
-def login(username="", password="", endpoint="https://app.dimensions.ai", instance="live", key="", verbose=True):
+def login(username="", password="", 
+            endpoint="https://app.dimensions.ai", 
+            instance="live", 
+            key="", 
+            verbose=True):
     """Login into the Dimensions API and store the query token in memory. 
 
     Two cases:
@@ -62,7 +66,17 @@ def login(username="", password="", endpoint="https://app.dimensions.ai", instan
         The API key (available to some users instead of username/password)
     verbose: bool, optional
         Verbose mode. Default: True.
-       
+
+
+    Notes
+    ---------------
+    The endpoint value can either be simply the Dimensions server hostname or the full API endpoint path. All the options below are valid endpoints: 
+
+        * `https://app.dimensions.ai`
+        * `https://app.dimensions.ai/api/dsl/v1` 
+        * `https://app.dimensions.ai/api/dsl/v2`
+
+
 
     Example
     -------
@@ -82,10 +96,13 @@ def login(username="", password="", endpoint="https://app.dimensions.ai", instan
     
     >>> dimcli.login(key="my-secret-key", endpoint="https://your-url.dimensions.ai")
 
+    See Also
+    ---------------
+    dimcli.core.api.Dsl
 
     """
 
-    from .core.auth import do_global_login, get_connection
+    from .core.auth import do_global_login, get_global_connection
 
     try:
         do_global_login(instance, username, password, key, endpoint)
@@ -93,28 +110,28 @@ def login(username="", password="", endpoint="https://app.dimensions.ai", instan
         print("Login failed: please ensure your credentials are correct.")
         raise(e)
 
-    CONNECTION = get_connection()
+    CONNECTION = get_global_connection()
 
-    if CONNECTION['token'] and verbose:
-        _print_login_success(CONNECTION, username, password, key)
+    if CONNECTION.token and verbose:
+        _print_login_success(username, password, key)
         print_dimcli_report_if_outdated()
 
 
 
-def _print_login_success(CONNECTION, username, password, key):
+def _print_login_success(username, password, key):
     click.secho("Dimcli - Dimensions API Client (" + VERSION + ")", dim=True)
     CLIENT = Dsl(verbose=False)
     # dynamically retrieve dsl version 
     try:
-        _info = CLIENT.query("describe version")['release']
+        _info = "v" + CLIENT.query("describe version")['release']
     except:
-        _info = "not available"
+        _info = "[failed to retrieve version information]"
 
     if (username and password) or key:
         _method = "manual login"
     else:
         _method = "dsl.ini file"
-    click.secho(f"Connected to: {CLIENT._url} - DSL v{_info}", dim=True)
+    click.secho(f"Connected to: <{CLIENT._url}> - DSL {_info}", dim=True)
     click.secho(f"Method: {_method}", dim=True)
 
 
@@ -131,10 +148,10 @@ def logout():
     >>> dimcli.logout()
 
     """
-    from .core.auth import reset_login, get_connection
-    CONNECTION = get_connection()
-    if CONNECTION['token']:
-        reset_login()
+    from .core.auth import get_global_connection
+    CONNECTION = get_global_connection()
+    if CONNECTION.token:
+        CONNECTION.reset_login()
         print("Logging out... done") 
     else:
         print("Please login first") 
@@ -156,11 +173,12 @@ def login_status():
     False
 
     """
-    from .core.auth import get_connection
-    CONNECTION = get_connection()
-    if CONNECTION['token']:
-        print("Dimcli %s - Succesfully connected to <%s>" % (str(VERSION), CONNECTION['url'])) 
+    from .core.auth import get_global_connection
+    CONNECTION = get_global_connection()
+    if CONNECTION.token:
+        print("Dimcli %s - Succesfully connected to <%s>" % (str(VERSION), CONNECTION.url)) 
         return True
     else:
         print("Status: not logged in") 
         return False
+